Test dijkstra_mat function in catch
Rewrite Dijkstra_mat as a wrapper function, move logic into
InternalDijkstra_mat. Test this latter function in Catch using the
testthat wrapper. Add a header file to declare this internal function
in both the test file and the code file.  Attempting to do this with
the Dijkstra_mat function (declaring the function elsewhere) causes it
to disappear from RcppExports.R, and breaks the R code.

Working on #2